# docs/swagger.yaml
basePath: /api/v1
definitions:
  delivery.AccessTokenResponse:
    properties:
      access_token:
        type: string
      email:
        type: string
      expires_at:
        type: string
      role:
        type: string
    type: object
...
  delivery.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
  delivery.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
      remember:
        type: boolean
    type: object
...
  entity.User:
    properties:
      country_code:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
host: gosm.muhammadilham.xyz
info:
  contact:
    email: hi@muhammadilham.xyz
    name: API Support
    url: https://pumbook.muhammadilham.xyz
  description: A guest management system that is written in Go.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GoSM
  version: "1.0"
paths:
  /api/v1/auth:
    post:
      consumes:
      - application/json
      parameters:
      - description: User sign-in credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/delivery.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully authenticated
          schema:
            allOf:
            - $ref: '#/definitions/delivery.Response'
            - properties:
                data:
                  $ref: '#/definitions/delivery.AccessTokenResponse'
              type: object
        "400":
          description: Invalid credentials or bad request
          schema:
            $ref: '#/definitions/delivery.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/delivery.Response'
      summary: Authenticate user and return access token
      tags:
      - auth
  /api/v1/auth/profile:
    get:
      consumes:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/delivery.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery.Response'
      summary: get logged user's profile.
      tags:
      - auth
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: User sign-up credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/delivery.SignInRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/delivery.Response'
            - properties:
                data:
                  $ref: '#/definitions/delivery.AccessTokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery.Response'
      summary: Register a new user
      tags:
      - auth
...
swagger: "2.0"